requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.15
django-cors-headers==4.4.0
django-filter==24.3
django-memcached==0.1.2
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
firebase-admin==6.5.0
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.8
inflection==0.5.1
msgpack==1.0.8
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pymemcache==4.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-memcached==1.62
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
